src/backend/pyproject.toml
[project]
name = "drive-organizer"
version = "0.1.0"
description = "OneDrive file organizer with multi-account Microsoft authentication"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "httpx>=0.28.0",
    "msal>=1.31.0",
    "pyjwt>=2.10.0",
    "cryptography>=44.0.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.22",
    "arq>=0.26.3",
    "aiosqlite>=0.22.1",
    "pypdf>=5.2.0",
    "pillow>=11.1.0",
    "rarfile>=4.2",
    "py7zr>=0.22.0",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.4",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.6",
    "ultralytics>=8.3.0",
    "open-clip-torch>=2.26.1",
    "rapidocr-onnxruntime>=1.4.4",
    "exifread>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
